updated before release

+++ app/Http/Controllers/Admin/AdminUserController.php
@@ -3,9 +3,12 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\Admin\AdminTableRequest;
 use App\Http\Requests\Admin\AdminUserRequest;
+use App\Http\Requests\ProfileUpdateRequest;
 use App\Http\Resources\Admin\AdminUserResource;
 use App\Models\Activity;
+use App\Models\Cart;
 use App\Models\User;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Http\Request;
@@ -23,17 +26,18 @@ class AdminUserController extends Controller
         if ($search_users) {
             $users = User::query()
                 ->where('name', 'LIKE', "%$search_users%")
-                ->select('id', 'name', 'username', 'email', 'number_phone', 'address', 'status', 'picture')
+                ->select('id', 'name', 'username', 'email', 'number_phone', 'address', 'status', 'picture', 'is_active')
                 ->where('status', 'employee')
                 ->latest()
                 ->fastPaginate(10)->withQueryString();
         } else {
             $users = User::query()
-                ->select('id', 'name', 'username', 'email', 'number_phone', 'address', 'status', 'picture')
+                ->select('id', 'name', 'username', 'email', 'number_phone', 'address', 'status', 'picture', 'is_active')
                 ->where('status', 'employee')
                 ->latest()
                 ->fastPaginate();
         }
+
         return inertia('Admin/Employee/Index', [
             "employees" => AdminUserResource::collection($users),
             "total_employees" => $total_employee,
@@ -51,6 +55,7 @@ class AdminUserController extends Controller
             "address" => $request->address,
             "status" => $request->status,
             "password" => Hash::make("randa091100"),
+            'is_active' => $request->is_active == 1 ? true : ($request->is_active == 0 ? false : null),
             "picture" => $request->hasFile('picture') ? $picture->storeAs('images/employees', $name . '.' . $picture->extension()) : null
         ]);
 
@@ -65,6 +70,12 @@ class AdminUserController extends Controller
     public function update(AdminUserRequest $request, User $user)
     {
         $picture = $request->file('picture');
+        if ($picture) {
+            $request->validate([
+                'picture' => ['nullable', 'mimes:png,jpg,jpeg', 'image', 'max:2048'],
+            ]);
+        }
+
         $user->update([
             "name" => $name = $request->name ?: $user->name,
             "username" => $request->username ?: $user->username,
@@ -72,7 +83,7 @@ class AdminUserController extends Controller
             "number_phone" => $request->number_phone ?: $user->number_phone,
             "address" => $request->address ?: $user->address,
             "status" => $request->status ? $request->status : $user->status,
-            "password" => $user->password,
+            'is_active' => $request->is_active == 1 ? true : ($request->is_active == 0 ? false : $user->is_active),
             "picture" => $request->hasFile('picture') ? $picture->storeAs('images/employees', $name . '.' . $picture->extension()) : $user->picture
         ]);
 
@@ -86,15 +97,22 @@ class AdminUserController extends Controller
 
     public function destroy(User $user)
     {
-        if ($user->picture) {
-            Storage::delete($user->picture);
+        if ($user->carts()->count() > 0 || $user->invoices()->count() > 0) {
+            $user->update([
+                "is_active" => false,
+            ]);
+            return false;
+        } else {
+            if ($user->picture) {
+                Storage::delete($user->picture);
+            }
+
+            Activity::create([
+                "activity" => Auth::user()->name . " Deleted Employee " . $user->name
+            ]);
+
+            $user->delete();
+            return back();
         }
-
-        Activity::create([
-            "activity" => Auth::user()->name . " Deleted Employee " . $user->name
-        ]);
-
-        $user->delete();
-        return back();
     }
 }
